코드 스타일 수정 및 Step-2 완성
