+++ PhotosApp/PhotosApp/ViewController.swift
@@ -18,19 +18,18 @@ class ViewController: UIViewController {
     //MARK: - Lifecycle
     override func viewDidLoad() {
         super.viewDidLoad()
-        
+        PHPhotoLibrary.shared().register(self)
         mainCollectionView.delegate = self
         mainCollectionView.dataSource = self
-        
-        let nibName = UINib(nibName: "CollectionViewCell", bundle: nil)
-        mainCollectionView.register(nibName, forCellWithReuseIdentifier: "CollectionViewCell")
+        mainCollectionView.register(CollectionViewCell.nib(), forCellWithReuseIdentifier: CollectionViewCell.identifier)
         requestImage()
         self.mainCollectionView.reloadData()
     }
     
+    //MARK: - fetchAssets 메소드
     func requestImage() {
         let fetchOptions = PHFetchOptions()
-        fetchOptions.sortDescriptors = [NSSortDescriptor(key: "creationDate", ascending: false)]
+        fetchOptions.sortDescriptors = [NSSortDescriptor(key: "creationDate", ascending: true)]
         self.allPhotos = PHAsset.fetchAssets(with: fetchOptions)
     }
 }
@@ -43,17 +42,13 @@ extension ViewController: UICollectionViewDataSource {
     }
     
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "CollectionViewCell", for: indexPath) as! CollectionViewCell
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: CollectionViewCell.identifier, for: indexPath) as! CollectionViewCell
         
         let asset: PHAsset = allPhotos!.object(at: indexPath.row)
         imageManager.requestImage(for: asset, targetSize: CGSize(width: 100, height: 100), contentMode: .aspectFill, options: nil, resultHandler: {image, _ in cell.cellImageView.image = image})
         return cell
     }
     
-    // 편집 가능한지 요청하는 메소드
-    func collectionView(_ collectionView: UICollectionView, canEditItemAt indexPath: IndexPath) -> Bool {
-        return true
-    }
 }
 
 extension ViewController: UICollectionViewDelegateFlowLayout {
@@ -73,7 +68,7 @@ extension ViewController: PHPhotoLibraryChangeObserver {
         if let changed = changeInstance.changeDetails(for: allPhotos!){
             allPhotos = changed.fetchResultAfterChanges
         }
-        DispatchQueue.main.sync {
+        DispatchQueue.main.async {
             self.mainCollectionView.reloadData()
         }
     }
